Move some common UI components to the `components/common`.
