+++ src/components/common/footer/Footer.container.ts
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import { RootState } from '../../store/state';
+import { RootState } from '../../../store/state';
 import Footer from './Footer';
 
 // eslint-disable-next-line no-unused-vars
