+++ src/components/common/footer/Footer.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import 'src/variables';
 
 .footer {
   display: flex;
