+++ src/gen/types/KeyboardDefinition.ts
@@ -10,7 +10,14 @@ export interface KeyboardDefinitionSchema {
   vendorId: string;
   productId: string;
   lighting?:
-    | ('none' | 'qmk_backlight' | 'qmk_rgblight' | 'qmk_backlight_rgblight' | 'wt_rgb_backlight' | 'wt_mono_backlight')
+    | (
+        | 'none'
+        | 'qmk_backlight'
+        | 'qmk_rgblight'
+        | 'qmk_backlight_rgblight'
+        | 'wt_rgb_backlight'
+        | 'wt_mono_backlight'
+      )
     | {
         extends?: string;
         effects?: [] | [[] | [string] | [string, number]];
